rename server directory
- fix string representation of models
- fix file path issues
