+++ a_socials/settings/base.py
@@ -40,7 +40,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'webpack_loader',
-    'andela_socials',
+    'api',
     'rest_framework',
     'django_filters',
 ]
@@ -69,7 +69,7 @@ TEMPLATES = [
                 'django.template.context_processors.request',
                 'django.contrib.auth.context_processors.auth',
                 'django.contrib.messages.context_processors.messages',
-                'andela_socials.context_processors.set_environment'
+                'api.context_processors.set_environment'
             ],
         },
     },
