+++ a_socials/urls.py
@@ -17,12 +17,12 @@ from django.conf.urls import url, include
 from django.contrib import admin
 
 from django.views.generic.base import TemplateView
-from andela_socials import views
+from api import views
 
 urlpatterns = [
     url(r'^$', views.HomeView.as_view(), name='homepage'),
     url(r'^admin/?', admin.site.urls),
-    url(r'^api/v1/?', include('andela_socials.urls')),
-    url(r'^/', include('andela_socials.urls')),
+    url(r'^api/v1/?', include('api.urls')),
+    url(r'^/', include('api.urls')),
     url(r'^.*/?', views.DashBoardView.as_view(template_name="index.html"), name='base')
 ]
